// src/types/project.ts
export interface Project {

  id: string;

  slug: string;

  title: string;

  subtitle: string;

  client: string;

  year: string;

  role: string;

  website: string;

  services: string[];

  image: string;

  coverImage: string;

  heroImage: string;

  overview: string;

  challenge: string;

  solution: string;

  technologies: string[];

  gallery: string[];

}